Use updated top and left when aligning tooltips.
This fixes bugs with tooltips that were being rendered partially offscreen. Tooltips position themselves based on content size and offset. Part of the alignment pass ensures content is onscreen after centering it against the tooltip target. There was a bug that prevented the right and bottom from using the 'normalized' left and top.